Refactoring pass over benchmark directory (#744)
