+++ src/CMakeLists.txt
@@ -12,7 +12,6 @@ endif()
 add_compile_options(-pthread -Wall -Wextra -Wpedantic -Werror)
 
 include_directories(
-    ${PROJECT_SOURCE_DIR}/src/benchmark/lib
     ${PROJECT_SOURCE_DIR}/src/lib
     ${PROJECT_SOURCE_DIR}/third_party/cxxopts/include
     ${PROJECT_SOURCE_DIR}/third_party/magic_enum/include
