+++ src/benchmark/CMakeLists.txt
@@ -1,2 +1,6 @@
+include_directories(
+    ${PROJECT_SOURCE_DIR}/src/benchmark/lib
+)
+
 add_subdirectory(bin)
 add_subdirectory(lib)
