+++ src/benchmark/lib/tpch/tpch_mock_catalog.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "metadata/mock_catalog.hpp"
+
+namespace skyrise {
+
+/**
+ * Provides the schema for all TPC-H tables.
+ */
+class TpchMockCatalog : public MockCatalog {
+ public:
+  TpchMockCatalog();
+};
+
+}  // namespace skyrise
