--- src/lib/metadata/tpch_mock_catalog.hpp
@@ -1,20 +1,0 @@
-#pragma once
-
-#include <memory>
-#include <string>
-#include <unordered_map>
-
-#include "mock_catalog.hpp"
-#include "table_schema.hpp"
-
-namespace skyrise {
-
-/**
- * Provides TableSchema data for all TPC-H tables.
- */
-class TpchMockCatalog : public MockCatalog {
- public:
-  TpchMockCatalog();
-};
-
-}  // namespace skyrise
