+++ src/test/benchmark/lib/ec2/ec2_benchmark_runner_test.cpp
@@ -5,7 +5,7 @@
 
 #include "client/client.hpp"
 #include "ec2/ec2_benchmark_config.hpp"
-#include "lib/testing/aws_test.hpp"
+#include "testing/aws_test.hpp"
 
 namespace skyrise {
