+++ src/test/benchmark/lib/lambda/lambda_benchmark_runner_test.cpp
@@ -5,7 +5,7 @@
 
 #include "client/client.hpp"
 #include "lambda/lambda_benchmark_config.hpp"
-#include "lib/testing/aws_test.hpp"
+#include "testing/aws_test.hpp"
 
 namespace skyrise {
