+++ src/test/lib/monitoring/function_segments_test.cpp
@@ -14,8 +14,8 @@
 
 #include "client/client.hpp"
 #include "function/function_utils.hpp"
-#include "lib/testing/aws_test.hpp"
 #include "monitoring_test_utils.hpp"
+#include "testing/aws_test.hpp"
 #include "utils/assert.hpp"
 #include "utils/string.hpp"
