+++ src/test/lib/monitoring/metrics_collector_test.cpp
@@ -12,8 +12,8 @@
 #include <gtest/gtest.h>
 
 #include "client/client.hpp"
-#include "lib/testing/aws_test.hpp"
 #include "monitoring_test_utils.hpp"
+#include "testing/aws_test.hpp"
 #include "utils/assert.hpp"
 
 namespace skyrise {
