+++ src/test/lib/utils/costs/cost_calculator_test.cpp
@@ -7,7 +7,7 @@
 #include <gtest/gtest.h>
 
 #include "client/client.hpp"
-#include "lib/testing/aws_test.hpp"
+#include "testing/aws_test.hpp"
 #include "utils/costs/pricing.hpp"
 #include "utils/unit_conversion.hpp"
